// src/app/models/mantenimiento.ts
export class Mantenimiento {
    id?: number;
    fechaInicio?: string;
    proximoMant?: string;
    fecha_fin?: string | null; 
    modelo?: string;
    marca?: string;
    detalles?: string;
    detalle2?: string;
    vehiculo?: number;
    placa?: string;
    estado?: string ;
    tipo_mantenimiento?: number;

    nombre?: string;
    taller?: number; 
}